Strafe by gyro correction stuffs
So, tons of stuff happened, The new playground program was to try to strafe while correcting, but it didn't work. Ritvik came and said that was wrong so edits to the ARYA program so the robot corrects after strafing as straight as possible. Shouldn't be too many bad this meet bc just have to check the robot's turn tendencies and values and stuff. So, yeah, GL w that